Move mesos test to examples/
