+++ integration/mesos/pom.xml
@@ -58,11 +58,6 @@
       <artifactId>alluxio-underfs-hdfs</artifactId>
       <version>${project.version}</version>
     </dependency>
-    <dependency>
-      <groupId>com.beust</groupId>
-      <artifactId>jcommander</artifactId>
-      <scope>test</scope>
-    </dependency>
 
     <!-- Other projects' test-jars -->
     <dependency>
